Menu Gui_Items now take in px size

+++ cpp/src/utils/button.h
@@ -6,7 +6,7 @@ class Button : public Gui_Item {
 	Text_Field button_text;
 
   public:
-	Button(float width, float height, float start_pos_x, float start_pos_y, std::string text)
+	Button(int width, int height, int start_pos_x, int start_pos_y, std::string text)
 		: Gui_Item(width, height, start_pos_x, start_pos_y),
 		  button_text(width, height, start_pos_x, start_pos_y, text) {}
